Добавлены модули, вненсены небольшие правки в служебные файлы проекта
