+++ .eslintrc.yml
@@ -1,4 +1,4 @@
-parserOption:
+parserOptions:
   ecmaVersion: 2015
   sourceType: 'module'
