+++ package.json
@@ -4,9 +4,9 @@
   "private": true,
   "description": "Личный проект «Киноман» от HTML Academy",
   "scripts": {
-    "build": "webpack",
-    "test": "npm run eslint",
     "eslint": "eslint src/**/*.js",
+    "test": "npm run eslint",
+    "build": "webpack",
     "start": "webpack-dev-server --open"
   },
   "devDependencies": {
