+++ public/index.html
@@ -34,7 +34,7 @@
       <!-- Количество фильмов -->
     </section>
   </footer>
-  <script src='js/main.js'></script>
+  <script src='./bundle.js'></script>
 </body>
 
 </html>
