+++ src/components/film-details.js
@@ -1,99 +1,5 @@
-'use strict';
-
-const FILMS_NUMER = 5;
-const SUB_FILMS_NUMBER = 2;
-
-// Блок рэйтинга с фото-профиля для шапки
-const createProfileTemplate = () => {
-  return (
-    `<section class="header__profile profile">
-    <p class="profile__rating">Movie Buff</p>
-    <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-    </section>`
-  )
-};
-
-// Блок меню включая статистику
-const createMainMenuTemplate = () => {
-  return (
-    `<nav class="main-navigation">
-    <div class="main-navigation__items">
-      <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
-      <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">13</span></a>
-      <a href="#history" class="main-navigation__item">History <span class="main-navigation__item-count">4</span></a>
-      <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">8</span></a>
-    </div>
-    <a href="#stats" class="main-navigation__additional">Stats</a>
-    </nav>`
-  )
-};
-
-// Блок сортировки
-const createSortMenuTemplate = () => {
-  return (
-    `<ul class="sort">
-      <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
-      <li><a href="#" class="sort__button">Sort by date</a></li>
-      <li><a href="#" class="sort__button">Sort by rating</a></li>
-    </ul>`
-  )
-};
-
-// Блок с фильмами, список фильмов и два экстра-блока
-const createFilmsListTemlate = () => {
-  return (
-    `<section class="films">
-      <section class="films-list">
-        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-        <div class="films-list__container">
-        </div>
-      </section>
-      <section class="films-list--extra">
-        <h2 class="films-list__title">Top rated</h2>
-        <div class="films-list__container">
-        </div>
-      </section>
-      <section class="films-list--extra">
-        <h2 class="films-list__title">Most commented</h2>
-        <div class="films-list__container">
-        </div>
-      </section>
-    </section>`
-  )
-}
-
-// Карточка фильма
-const createFilmCardTemlate = () => {
-  return (
-    `<article class="film-card">
-          <h3 class="film-card__title">The Dance of Life</h3>
-          <p class="film-card__rating">8.3</p>
-          <p class="film-card__info">
-            <span class="film-card__year">1929</span>
-            <span class="film-card__duration">1h 55m</span>
-            <span class="film-card__genre">Musical</span>
-          </p>
-          <img src="./images/posters/the-dance-of-life.jpg" alt="" class="film-card__poster">
-          <p class="film-card__description">Burlesque comic Ralph "Skid" Johnson (Skelly), and specialty dancer Bonny Lee King (Carroll), end up together on a cold, rainy night at a tr…</p>
-          <a class="film-card__comments">5 comments</a>
-          <form class="film-card__controls">
-            <button class="film-card__controls-item button film-card__controls-item--add-to-watchlist">Add to watchlist</button>
-            <button class="film-card__controls-item button film-card__controls-item--mark-as-watched">Mark as watched</button>
-            <button class="film-card__controls-item button film-card__controls-item--favorite">Mark as favorite</button>
-          </form>
-    </article>`
-  )
-};
-
-// Кнопка показать больше
-const createShowMoreButtonTemlate = () => {
-  return (
-    `<button class="films-list__show-more">Show more</button>`
-  )
-};
-
 // Карточка с деталями фильма
-const createFilmDeatilsTemplate = () => {
+export const createFilmDeatilsTemplate = () => {
   return (
     `<section class="film-details">
       <form class="film-details__inner" action="" method="get">
@@ -264,57 +170,5 @@ const createFilmDeatilsTemplate = () => {
         </div>
       </form>
     </section>`
-  )
+  );
 };
-
-// Функция вставки объекта на страницу
-const renderBlock = (place, block, position = 'beforeend') => {
-  place.insertAdjacentHTML(position, block);
-};
-
-// Шапка страницы
-const pageHeader = document.querySelector('.header');
-// Блок main
-const pageMain = document.querySelector('.main');
-
-// Отрисовка элементов на странице
-renderBlock(pageHeader, createProfileTemplate());
-renderBlock(pageMain, createMainMenuTemplate());
-renderBlock(pageMain, createSortMenuTemplate());
-renderBlock(pageMain, createFilmsListTemlate());
-
-// Блок с фильмами
-const filmsSection = document.querySelector('.films');
-// Основной блок с фильмами
-const mainFilmsBlock = filmsSection.querySelector('.films-list');
-// Список всех 3х блоков
-const filmsLists = filmsSection.querySelectorAll('.films-list__container');
-// Завденеие блоков в отдельные переменыне
-const [mainFilmList, topRatedFilms, mostCommentedFilms] = filmsLists;
-
-// Отрисовка основных фильмов
-for (let i = 0; i < FILMS_NUMER; i++) {
-  renderBlock(mainFilmList, createFilmCardTemlate());
-}
-
-// Отрисовка кнпоки показать еще
-renderBlock(mainFilmsBlock, createShowMoreButtonTemlate());
-
-// Отрисовка просматриваемых и комментируемых фильмов
-for (let i = 0; i < SUB_FILMS_NUMBER; i++) {
-  renderBlock(topRatedFilms, createFilmCardTemlate());
-  renderBlock(mostCommentedFilms, createFilmCardTemlate());
-}
-
-const pageFooter = document.querySelector('.footer');
-// Отрисовка карточки фильма
-renderBlock(pageFooter, createFilmDeatilsTemplate(), 'afterend');
-
-// Временный блок
-const popup = document.querySelector('.film-details');
-const popupCloseButton = popup.querySelector('.film-details__close-btn');
-const onPopupClose = () => {
-  popup.remove();
-  popupCloseButton.removeEventListener('click', onPopupClose);
-};
-popupCloseButton.addEventListener('click', onPopupClose);
