+++ src/components/show-more-button.js
@@ -1,0 +1,6 @@
+// Кнопка показать больше
+export const createShowMoreButtonTemlate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
