+++ webpack.config.js
@@ -1,13 +1,13 @@
 const path = require('path');
 
-module.export = {
+module.exports = {
   mode: 'development',
-  entry: '.src/main.js',
+  entry: './src/main.js',
   output: {
     filename: 'bundle.js',
     path: path.join(__dirname, 'public'),
   },
-  devtools: 'source-map',
+  devtool: 'source-map',
   devServer: {
     contentBase: path.join(__dirname, 'public'),
     watchContentBase: true,
